Registrationpage

+++ roomranger/src/main/java/org/launchcode/roomranger/data/UserRepository.java
@@ -4,7 +4,9 @@ import org.launchcode.roomranger.models.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
-//    User findByUserName(String username);
+public interface UserRepository extends CrudRepository<User, Long> {
+      // User findByUsername(String username);
 }
